requirements.txt
Flask
flask-PyMongo
python-dotenv
passlib
markupsafe 